third commit; completed exercise but not further study

+++ app.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template, redirect, flash, session
-from models import db, connect_db, User
+from models import db, connect_db, User, Feedback
 from forms import RegistrationForm, LoginForm, FeedbackForm
 
 
@@ -18,9 +18,14 @@ db.create_all()
 def index():
     """Index."""
 
-    return redirect("/register")
+    if "username" in session:
+        username = session["username"]
+        user_info = User.query.get_or_404(username)
+        return redirect(f"/users/{user_info.username}")
+    else:
+        return redirect("/register")
 
-
+# REGISTER NEW USER
 @app.route("/register", methods=["GET", "POST"])
 def register_user():
     """Page to register new user."""
@@ -43,12 +48,15 @@ def register_user():
 
         db.session.commit()
 
+        session["username"] = new_user.username
+
         flash(f"Congrats, {first_name}! You have succesfully registered!")
-        return redirect(f"/users/{new_user.username}")  # USER'S HOMEPAGE, ONCE LOGGED IN
+        # USER'S HOMEPAGE, ONCE LOGGED IN
+        return redirect(f"/users/{new_user.username}")
     else:
         return render_template("register.html", form=form)
 
-
+# USER LOGIN
 @app.route("/login", methods=["GET", "POST"])
 def login_user():
     """Page for user to login."""
@@ -69,19 +77,19 @@ def login_user():
 
     return render_template("login.html", form=form)
 
-
+# USER HOMEPAGE
 @app.route("/users/<username>")
 def user_page(username):
     """ displays user's homepage. """
 
     if "username" not in session:
         flash("You must login first!")
-        return redirect("/")
+        return redirect("/")  # REDIRECT TO LOGIN PAGE
     else:
         user_info = User.query.get_or_404(username)
         return render_template("user.html", user_info=user_info)
 
-
+# LOGOUT
 @app.route("/logout")
 def logout():
     """Logs user out and redirects to home."""
@@ -90,7 +98,7 @@ def logout():
 
     return redirect("/")
 
-
+# DELETE USER
 @app.route("/users/<username>/delete")
 def delete_user(username):
     """ deletes a user. """
@@ -99,13 +107,18 @@ def delete_user(username):
         flash("You must login first!")
         return redirect("/")
     else:
-        user_info = User.query.get_or_404(username)
-        db.session.delete(user_info)
-        db.session.commit()
+        if session["username"] == username:
+            user_info = User.query.get_or_404(username)
+            db.session.delete(user_info)
+            db.session.commit()
 
-        flash("User has been deleted")
-        return redirect("/")
+            flash("User has been deleted")
+            return redirect("/")
+        else:
+            flash("THAT IS NOT ALLOWED!")
+            return redirect("/")
 
+# ADD FEEDBACK
 @app.route("/users/<username>/feedback/add", methods=["GET", "POST"])
 def add_feedback(username):
     """ adds feedback. """
@@ -122,10 +135,46 @@ def add_feedback(username):
             title = form.title.data
             content = form.content.data
 
-            new_feedback = Feedback(title=title, content=content)
+            new_feedback = Feedback(
+                title=title, content=content, giver_name=session["username"], receiver_name=user_info.username)
             db.session.add(new_feedback)
             db.session.commit()
-            
+
             return redirect(f"/users/{user_info.username}")
         else:
             return render_template("feedback.html", form=form)
+
+# UPDATE FEEDBACK
+@app.route("/feedback/<int:feedback_id>/update", methods=["GET", "POST"])
+def update_feedback(feedback_id):
+    """Form to update existing feedback."""
+
+    feedback_info = Feedback.query.filter_by(id=feedback_id).first()
+
+    form = FeedbackForm(obj=feedback_info)
+
+    if "username" not in session:
+        flash("You must login first!")
+        return redirect("/")
+    elif session["username"] == feedback_info.giver_name and form.validate_on_submit():
+        feedback_info.title = form.title.data
+        feedback_info.content = form.content.data
+        db.session.commit()
+
+        return redirect(f"/users/{feedback_info.receiver_name}")
+    else:
+        return render_template("edit-feedback.html",
+                               form=form,
+                               feedback_info=feedback_info)
+
+# DELETE FEEDBACK
+@app.route("/feedback/<int:feedback_id>/delete", methods=["POST"])
+def delete_feedback(feedback_id):
+    """Delete a particular feedback."""
+    feedback_info = Feedback.query.filter_by(id=feedback_id).first()
+
+    if session["username"] == feedback_info.giver_name:
+        db.session.delete(feedback_info)
+        db.session.commit()
+
+        return redirect(f"/users/{feedback_info.receiver_name}")
